Task: Created Statistics functions Task: Test Cases created. Update: Settings and configurations updated.

+++ Calculator/Division.py
@@ -1,8 +1,7 @@
 def division(value1, value2):
     try:
-        value1 = float(value1)
-        value2 = float(value2)
+        value1 = int(value1)
+        value2 = int(value2)
         return float(value2 / value1)
     except:
         print("Divide by zero not possible.")
-
