+++ Calculator/Subtraction.py
@@ -1,4 +1,4 @@
 def subtraction(value1, value2):
-    value1 = float(value1)
-    value2 = float(value2)
-    return float(value1 - value2)
+    value1 = int(value1)
+    value2 = int(value2)
+    return value1 - value2
